Debug collate

+++ realworld_benchmark/nets/molecules_graph_regression/eig_net.py
@@ -101,5 +101,5 @@ class EIGNet(nn.Module):
         return self.MLP_layer(hg)
 
     def loss(self, scores, targets):
-        loss = nn.L1Loss()(scores, targets)
+        loss = nn.L1Loss()(scores, torch.FloatTensor(targets))
         return loss
